wip: start working on migrations
