--- core/src/db/mod.rs
@@ -1,5 +1,0 @@
-mod client;
-mod common;
-
-pub use client::create_connection;
-pub use common::{CountQueryReturn, JournalMode, JournalModeQueryResult};
